add widget logic

+++ src/main/java/ru/esina/widget/repository/WidgetRepository.java
@@ -16,5 +16,5 @@ public interface WidgetRepository {
 
     List<Widget> getAllWidget();
 
-    Double generateZCoordinate();
+    Long generateZCoordinate();
 }
