+++ src/main/java/ru/esina/widget/service/WidgetService.java
@@ -6,9 +6,9 @@ import java.util.UUID;
 import ru.esina.widget.model.Widget;
 
 public interface WidgetService {
-    Widget createWidget(Double x, Double y, Double z, Double width, Double height);
+    Widget createWidget(Double x, Double y, Long z, Double width, Double height);
 
-    Widget modifyWidget(UUID uuid, Double x, Double y, Double z, Double width, Double height);
+    Widget modifyWidget(UUID uuid, Double x, Double y, Long z, Double width, Double height);
 
     void deleteWidget(UUID uuid);
